refactor: move hot styles to their own module
