+++ source/main.js
@@ -1,19 +1,4 @@
-if (process.env.NODE_ENV !== 'production') {
-    const cb = window.webpackHotUpdate;
-
-    window.webpackHotUpdate = (...args) => {
-        const links = document.querySelectorAll('link[rel=stylesheet]');
-
-        for (let link of links) {
-            let { href } = link;
-            link.href = 'about:blank';
-            link.href = href;
-        }
-
-        return cb(...args);
-    }
-}
-
+import './lib/style';
 import './main.scss';
 
 const canvas = document.getElementById('js-canvas');
